Separate code for run
