+++ piperider_cli/event/__init__.py
@@ -4,7 +4,7 @@ import uuid
 from ruamel import yaml
 
 from piperider_cli.event.collector import Collector
-from piperider_cli.workspace import Configuration
+from piperider_cli.configuration import Configuration
 
 PIPERIDER_USER_HOME = os.path.expanduser('~/.piperider')
 PIPERIDER_USER_PROFILE = os.path.join(PIPERIDER_USER_HOME, 'profile.yml')
